naming convention, schema by domain and got rid of pullOutRest in resolvers
