+++ server/schema/index.js
@@ -1,7 +1,5 @@
-const UserSchema = require("./sUser");
-const ChatSchema = require("./sChat");
-const MessageSchema = require("./sMessage");
-const Queries = require("./queries");
-const Mutations = require("./mutations");
+const UserSchema = require("./User");
+const ChatSchema = require("./Chat");
+const MessageSchema = require("./Message");
 
-module.exports = [UserSchema, ChatSchema, MessageSchema, Queries, Mutations];
+module.exports = [UserSchema, ChatSchema, MessageSchema];
