--- server/schema/mutations.js
@@ -1,14 +1,0 @@
-const { gql } = require("apollo-server");
-
-const typeDefs = gql`
-  type Mutation {
-    createUser(input: CreateUserInput!): User!
-
-    createChat(input: CreateChatInput!): Chat!
-    addUsersToChat(input: AddUsersToChatInput!): Chat!
-
-    createMessage(input: CreateMessageInput!): Message!
-  }
-`;
-
-module.exports = typeDefs;
